pyproject.toml
[project]
name = "emotion-speech-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.65.0",
    "hydra-core>=1.3.2",
    "kagglehub>=0.3.13",
    "mlflow>=3.8.0",
    "numpy>=2.4.0",
    "omegaconf>=2.3.0",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "onnxscript>=0.5.7",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
    "soundfile>=0.13.1",
    "tensorrt>=10.14.1.48.post1",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
